+++ jsconfig.json
@@ -8,6 +8,7 @@
       "@components/*": ["./components/*"],
       "@services/*": ["./services/*"],
       "@hooks/*": ["./hooks/*"],
+      "@classes/*": ["./classes/*"],
     }
   },
   "exclude": ["node_modules", "dist"],
